Disable tensorflow due to upstream problem
We have just added tensorflow again, as the repository package was
fixed, allowing to properly use it here again. Unfortunately, there
is a tensorflow upstream issue[1] that prevents applications to
startup when they are linked to both ffmpeg and protobuf. We better
remove tensorflow again to not break client applications, until
this is fixed by upstream.

References
----------
[1] https://github.com/tensorflow/tensorflow/issues/59081#issuecomment-2093988476