Switch the gcc 14 fix to use a compiler flag
The flite audio filter is giving a segmentation fault, so we
better use only a compiler flag to fix the build with gcc 14.

When using just the compiler flag, the flite filter also gives
the segmentation fault, but in this circunstance it looks better
to leave the upstream source unthouched.