Switch libplacebo dependency to libplacebo-git. Fix svt patches.
This fixes a build error due to a recent change[1] in upstream
FFmpeg, which removed 'av_stream_get_side_data()'. It is fixed[2]
by upstream libplacebo, and the fix is present only in the git
master branch at the time of writing. Using libplacebo-git while
there is no stable libplacebo with this fix.

The original and now unmaintained svt-hevc and svt-vp9 patches
no longer applies on top of FFmpeg git master. Added fixes for
them.

[1] https://git.ffmpeg.org/gitweb/ffmpeg.git/commit/ec8e796b42bccf999fdf31c110385e2cc119b7f2
[2] https://github.com/haasn/libplacebo/commit/563ea4a1a23237c0acd9f0ba29f5166ebcd5cf60