Match latest upstream changes. Refresh patches.
Upstream splitted the C and C++ flags. Now, for correctly building
with decklink-sdk (which is C++ in FFmpeg), the include flag needs
to be also in CXXFLAGS, otherwise, a build error will happen at
compile time. The flag is still needed in CFLAGS for detecting the
decklink-sdk, otherwise, a build error will happen at configure
time.

Also changed the priority of the custom pkg-config search path in
this commit.

https://git.ffmpeg.org/gitweb/ffmpeg.git/commit/0ce413af9cc8040095a6b714600d6a8dceaca514